+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project, at least loosely, adheres to [Semantic Versioning](https://sem
 - Color-by-jump mode for pintk
 - `pytest-xdist` now allows `pytest -n auto` to use all cores on the machine to run tests in parallel; `make test` now does this.
 - Added the ability to write clock files in TEMPO or TEMPO2 format
+- Added examples of how to write a par file to tutorials
+- Added `TimingModel.write_parfile()`
 ### Fixed
 - INCLUDE lines in tim files are now relative to the location of the tim file (bug #1269)
 ### Changed